Mark as dead.
origami-migration-guru started as a 10% time project. It was
useful during the [2019 major
cascade](https://origami.ft.com/blog/2019/10/31/major-cascade/)
but is not easy to use, for example requiring a list of all
manifest files, and is not 100% reliable. It needs more work and
has therefore been marked as dead.

Note "deprecated" is mentioned in the README because "dead" sounds
so... harsh.